feat(core): optimize the merge method

+++ src-tauri/src/command/m3u8/merge.rs
@@ -3,18 +3,28 @@ use std::io::Read;
 use std::io::Write;
 use std::path::PathBuf;
 
-pub fn merge_ts(id_range: usize, temp_dir: &str) {
+pub fn merge_ts(temp_dir: &str) {
+    // 读取目录下所有文件路径
+    // 可能包含其他非ts文件 如.DS_Store
+    let mut paths = std::fs::read_dir(temp_dir)
+        .unwrap()
+        .map(|res| res.map(|e| e.path()))
+        .collect::<Result<Vec<_>, std::io::Error>>()
+        .unwrap();
+    paths.sort();
+
     let mut buffer = Vec::new();
     let out_path = PathBuf::from(temp_dir);
     let mut out_path = PathBuf::from(out_path.parent().unwrap());
     out_path.push("output.ts");
 
-    for id in 0..id_range {
+    for id in 0..paths.len() {
         let mut ts_path = PathBuf::from(temp_dir);
         ts_path.push(format!("{}.ts", id));
-        let mut f = File::open(&ts_path).unwrap();
-        f.read_to_end(&mut buffer).unwrap();
-        std::fs::remove_file(&ts_path).unwrap();
+        if let Ok(mut f) = File::open(&ts_path) {
+            f.read_to_end(&mut buffer).unwrap();
+            std::fs::remove_file(&ts_path).unwrap();
+        }
     }
 
     let mut f = File::create(&out_path).unwrap();
